Fixed the data types for Ids to match.

+++ src/main/java/com/bonginkosi/salaryservice/repository/SalaryRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SalaryRepository extends JpaRepository<Salary,Long> {
+public interface SalaryRepository extends JpaRepository<Salary,Integer> {
 
 
 }
